# Conflicts:
#	Snowy-Path/Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
#	Snowy-Path/Assets/Scenes/Dev/QB/03-Weather/QBWeather_Display.cs
#	Snowy-Path/Assets/Scripts/Character/Temperature.cs
#	Snowy-Path/Assets/Scripts/Character/Weather.cs
#	Snowy-Path/Assets/Scripts/Environment/WeatherZone.cs
#	Snowy-Path/fmod_editor.log

+++ Snowy-Path/Assets/Scripts/Character/Temperature.cs
@@ -128,7 +128,7 @@ public class Temperature : MonoBehaviour
         if (cloth != null)
             clothType = (float)cloth.type;
 
-        return (1 + m_weather.CurrentWeather.blizzardStrength) - ((1 + m_weather.CurrentWeather.blizzardStrength) * clothType / 100f);
+        return (1 + (float)m_weather.CurrentWeather.blizzardStrength) - ((1 + (float)m_weather.CurrentWeather.blizzardStrength) * clothType / 100f);
     }
 
     void OnTriggerEnter(Collider other)
